+++ IngameUserManager.cpp
@@ -8,7 +8,7 @@ void InGameUserManager::Init(uint16_t maxClientCount_) {
 	}
 }
 
-void InGameUserManager::Set(uint16_t connObjNum_, std::string userUuid_, std::string userId_, uint32_t userPk_, unsigned int userExp_, uint8_t userLevel_) {
+void InGameUserManager::Set(uint16_t connObjNum_, std::string userUuid_, std::string userId_, uint32_t userPk_, unsigned int userExp_, uint16_t userLevel_) {
 	inGmaeUsers[connObjNum_]->Set(userUuid_, userId_,userPk_, userExp_,userLevel_);
 }
